Add more UI utils -
This will need a redesign. There's a ton of imperative function calling, and the p5 sketch is being injected as an argument in a ton of spots.

Anything that requires the `p` argument should be extracted into a
class that has its own reference to the p object.

BUT! That can come later. it'll make it easier to instantiate/setup a sketch once it's implemented